pulling the canvas floating toolbars into their own file
